D2DRenderer 비트맵 부분 수정 - CreateD2DBitmapFromFile()시 이중 포인터로 반환하도록 변경 - CreateD2DBitmap() 중복 체크 통과시 해당 함수로 실제로 생성 - InitDirect2D() 발생시 모든 비트맵 재생성한다. - ReloadAllID2D1Bitmap() 모든 비트맵 재생성 하는 함수. 원래 있던 참조 이중 포인터들의 대상만 바꾼다. - ReleaseAllID2D1Bitmap() WinGameApp 루프 종료후 Uninitialize() 에서만 실행하도록 변경.
SpriteRenderer 변경점
- static 인스턴스 리스트 삭제 (필요 없어짐)
- ReloadImage 삭제 (D2DRenderer에서 직접 관리함)
- Image 맴버가 이중 포인터로 변경됨에 따른 Render 함수 수정.

SpriteAnimation 변경점
- AnimationAsset
-- Image 맴버가 이중 포인터로 변경됨에 따른 CreateAnimationClipFromFile() 함수 변경

+++ D2DEngine/Source/Vector/Vector2.cpp
@@ -149,7 +149,7 @@ Vector2 Vector2::BezierCurve(const Vector2& startPos, const Vector2& middlePos, 
 	return bezierCurve;
 }
 
-Vector2 Vector2::LagranCurve(float x, Vector2* pointArray, float pointArraySize)
+Vector2 Vector2::LagranCurve(float x, Vector2* pointArray, int pointArraySize)
 {
 	float* L = new float[pointArraySize] {};
 	for (int i = 0; i < pointArraySize; i++)
